makeitcamp rails 6.5

+++ db/migrate/20170728004813_add_price_per_night_to_rooms.rb
@@ -1,5 +1,5 @@
 class AddPricePerNightToRooms < ActiveRecord::Migration[5.0]
   def change
-    add_column :rooms, :price_per_night, :float
+    add_column :rooms, :price_per_night, :decimal
   end
 end
